+++ src/Ycs/Utils/ID.cs
@@ -20,9 +20,9 @@ namespace Ycs
         /// <summary>
         /// Unique per client id, continuous number.
         /// </summary>
-        public int Clock;
+        public long Clock;
 
-        public ID(long client, int clock)
+        public ID(long client, long clock)
         {
             Debug.Assert(client >= 0, "Client should not be negative, as it causes client encoder to fail");
             Debug.Assert(clock >= 0);
@@ -49,8 +49,8 @@ namespace Ycs
 
         public static ID Read(Stream reader)
         {
-            var client = (long)reader.ReadVarUint();
-            var clock = (int)reader.ReadVarUint();
+            var client = reader.ReadVarUint();
+            var clock = reader.ReadVarUint();
             Debug.Assert(client >= 0 && clock >= 0);
             return new ID(client, clock);
         }
